Send all removed objects within ack to current frame window
If object existed only for a single frame there is a non-zero probability for
client to miss update where it's removed. This may happen due to missed packet
with ack from client to server. Server makes updates based on diff between
world state in 2 given frames. If object was added in frame n and removed in
frame n + 1 but client got updates only on diffs between [n - a, n] and
[n - b, n + c] (a, b, c > 0) then object addition will be received but removal
will never be in any received updates.

The solution is to send all removed objects within range [n - a, n + b] for
each update based on such diff.